refactor component names , put home page inside home folder
