+++ src/components/layout/header/Header.jsx
@@ -5,10 +5,10 @@ import { Container } from "@mui/system";
 import Toolbar from "@mui/material/Toolbar";
 import logo from "../../../assets/HedearLogo/home-logo.svg";
 import useMediaQuery from "@mui/material/useMediaQuery";
-import BurgerMenu from "./menu/menu";
-import HeaderIcons from "./headerIcons/headerIcons";
-import SearchBar from "./searchBar/searchBar";
-import NavCategory from "./navCategory/navCategory";
+import BurgerMenu from "./Menu/Menu";
+import HeaderIcons from "./HeaderIcons/HeaderIcons";
+import SearchBar from "./SearchBar/SearchBar";
+import NavCategory from "./NavCategory/NavCategory";
 
 export default function Header() {
   const theme = useTheme();
