+++ src/components/layout/header/menu/menu.jsx
@@ -4,7 +4,7 @@ import Button from "@mui/material/Button";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
 import Fade from "@mui/material/Fade";
-import SearchBar from "../searchBar/searchBar";
+import SearchBar from "../SearchBar/SearchBar";
 import { FaRegRectangleXmark } from "react-icons/fa6";
 import { Box } from "@mui/system";
